+++ docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog for Typopo
 
 ## 2.1.3 // 2017-04-
+### Improvements
+* Fix abbreviations (e.g., i.e., a.m., p.m.), even when there are multiple spaces between them (e.    g. → e.g.)
+
 
 ### Fixes
 * fix Abbreviation module (it was falsely removing spaces at the beginning of the abbreviation)